feat: adiciona verificacao de funcionario com usuario vinculado, excluindo ambos

+++ microsservicos/identidade-acesso/src/main/java/com/projetos/manutencao/identidade_acesso/service/impl/FuncionarioServiceImpl.java
@@ -2,10 +2,11 @@ package com.projetos.manutencao.identidade_acesso.service.impl;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 import com.projetos.manutencao.identidade_acesso.dto.auth.FuncionarioDTO;
-import com.projetos.manutencao.identidade_acesso.model.Usuario;
 import com.projetos.manutencao.identidade_acesso.repository.UsuarioRepository;
+import com.projetos.manutencao.identidade_acesso.service.UsuarioService;
 import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,21 +20,19 @@ import com.projetos.manutencao.identidade_acesso.service.FuncionarioService;
 public class FuncionarioServiceImpl implements FuncionarioService {
     
     private final FuncionarioRepository repository;
-    private final UsuarioRepository usuarioRepository;
     @Autowired
     private ModelMapper modelMapper;
 
 
-    public FuncionarioServiceImpl(FuncionarioRepository repository, UsuarioRepository usuarioRepository) {
+    public FuncionarioServiceImpl(FuncionarioRepository repository) {
         this.repository = repository;
-        this.usuarioRepository = usuarioRepository;
     }
 
     public List<Funcionario> findAll() {
         return repository.findAll();
     }
 
-    public Optional<Funcionario> findById(Long id) {
+    public Optional<Funcionario> findById(String id) {
         return repository.findById(id);
     }
 
@@ -42,15 +41,18 @@ public class FuncionarioServiceImpl implements FuncionarioService {
         funcionario.setCargo(funcionarioDTO.getCargo());
         funcionario.setEquipe(funcionarioDTO.getCargo());
         funcionario.setEspecialidades(funcionarioDTO.getEspecialidades());
-        funcionario.setNome(funcionarioDTO.getCargo());
+        funcionario.setNome(funcionarioDTO.getNome());
         funcionario.setStatus(funcionarioDTO.getStatus());
 
         repository.save(funcionario);
         return repository.save(funcionario);
     }
 
-    public void deleteById(Long id) {
-        repository.deleteById(id);
+    public void deleteById(String id) {
+        Optional<Funcionario> funcionario = repository.findById(id.toString());
+        if (funcionario.isPresent()) {
+            repository.deleteById(id);
+        }
     }
 
     @Override
@@ -68,4 +70,29 @@ public class FuncionarioServiceImpl implements FuncionarioService {
 
         repository.save(funcionario);
     }
+
+    @Override
+    public Optional<Funcionario> findByMatricula(String matricula) {
+        Optional<Funcionario> funcionario = Optional.ofNullable(repository.findByMatricula(matricula)
+                .orElseThrow(() -> new EntityNotFoundException(
+                        "Funcionário não encontrado com a matrícula: " + matricula
+                )));
+
+        return funcionario;
+    }
+
+    public void vincularUsuarioAFuncionario(Funcionario funcionarioComUsuario){
+        repository.save(funcionarioComUsuario);
+    }
+
+    @Override
+    public Optional<Funcionario> findByUsuarioID(String usuarioId) {
+        Optional<Funcionario> funcionario = Optional.ofNullable(repository.findByUsuarioId(usuarioId)
+                .orElseThrow(() -> new EntityNotFoundException(
+                        "Funcionário não encontrado"
+                )));
+
+        return funcionario;
+    }
+
 }
